Fixup file naming for CommandLineTool, remove some cruft
