+++ cwl/src/main/scala/cwl/CommandLineTool.scala
@@ -1,66 +1,191 @@
 package cwl
 
-import eu.timepit.refined._
-import shapeless.{:+:, CNil}
-import cwl.CwlType._
+import common.Checked
+import common.validation.Checked._
+import shapeless.syntax.singleton._
+import shapeless.{:+:, CNil, Witness, Coproduct}
+import CommandLineTool._
+import CwlType.CwlType
+import CwlVersion._
+import eu.timepit.refined.W
+import wom.callable.Callable.{OutputDefinition, RequiredInputDefinition}
+import wom.callable.{Callable, TaskDefinition}
+import wom.executable.Executable
+import wom.expression.WomExpression
+import wom.types.WomFileType
+import wom.{CommandPart, RuntimeAttributes}
 
-case class CommandInputParameter(
-                                  id: String,
-                                  label: Option[String] = None,
-                                  secondaryFiles: Option[Array[Expression :+: String :+: CNil]] = None,
-                                  format: Option[Expression :+: Array[String] :+: String :+: CNil] = None, //only valid when type: File
-                                  streamable: Option[Boolean] = None, //only valid when type: File
-                                  doc: Option[String :+: Array[String] :+: CNil] = None,
-                                  inputBinding: Option[CommandLineBinding] = None,
-                                  default: Option[CwlAny] = None,
-                                  `type`: Option[MyriadInputType] = None)
+/**
+  * @param `class` This _should_ always be "CommandLineTool," however the spec does not -er- specify this.
+  */
+case class CommandLineTool private(
+                                   inputs: Array[CommandInputParameter],
+                                   outputs: Array[CommandOutputParameter],
+                                   `class`: Witness.`"CommandLineTool"`.T,
+                                   id: String,
+                                   requirements: Option[Array[Requirement]],
 
-case class CommandInputRecordSchema(
-                                     `type`: W.`"record"`.T,
-                                     fields: Option[Array[CommandInputRecordField]],
-                                     label: Option[String])
+                                   //TODO: Fix this when CwlAny parses correctly
+                                   //hints: Option[Array[CwlAny]] = None,
+                                   hints: Option[Array[Map[String, String]]],
 
-case class CommandInputRecordField(
-                                    name: String,
-                                    `type`: MyriadInputType,
-                                    doc: Option[String],
-                                    inputBinding: Option[CommandLineBinding],
-                                    label: Option[String])
-
-case class CommandInputEnumSchema(
-                                   symbols: Array[String],
-                                   `type`: W.`"enum"`.T,
                                    label: Option[String],
-                                   inputBinding: Option[CommandLineBinding])
+                                   doc: Option[String],
+                                   cwlVersion: Option[CwlVersion],
+                                   baseCommand: Option[BaseCommand],
+                                   arguments: Option[Array[CommandLineTool.Argument]],
+                                   stdin: Option[StringOrExpression],
+                                   stderr: Option[StringOrExpression],
+                                   stdout: Option[StringOrExpression],
+                                   successCodes: Option[Array[Int]],
+                                   temporaryFailCodes: Option[Array[Int]],
+                                   permanentFailCodes: Option[Array[Int]]) {
 
-case class CommandInputArraySchema(
-                                    items:
+  def womExecutable(inputFile: Option[String] = None): Checked[Executable] =
+    CwlExecutableValidation.buildWomExecutable(taskDefinition.validNelCheck, inputFile)
+
+  def taskDefinition: TaskDefinition = {
+
+    val id = this.id
+
+    val commandTemplate: Seq[CommandPart] = baseCommand.toSeq.flatMap(_.fold(BaseCommandToCommandParts)) ++
+      arguments.toSeq.flatMap(_.map(_.fold(ArgumentToCommandPart)))
+
+    val runtimeAttributes: RuntimeAttributes = RuntimeAttributes(Map.empty[String, WomExpression])
+
+    val meta: Map[String, String] = Map.empty
+    val parameterMeta: Map[String, String] = Map.empty
+
+    /*
+    quoted from: http://www.commonwl.org/v1.0/CommandLineTool.html#CommandOutputBinding :
+
+    For inputs and outputs, we only keep the variable name in the definition
+
+    The output parameter value is generated by applying these operations in the following order:
+
+      glob
+      loadContents
+      outputEval
+      secondaryFiles
+     */
+
+    val inputNames = this.inputs.map(_.id).toSet
+
+    val outputs: List[Callable.OutputDefinition] = this.outputs.map {
+      case CommandOutputParameter(cop_id, _, _, _, _, _, Some(outputBinding), Some(outputType)) if outputType.select[CwlType].contains(CwlType.File) =>
+        OutputDefinition(FullyQualifiedName(cop_id).id, WomFileType, CommandOutputExpression(outputBinding, WomFileType, inputNames))
+      case CommandOutputParameter(cop_id, _, _, _, _, _, Some(outputBinding), Some(tpe)) =>
+        val womType = tpe.select[CwlType].map(cwlTypeToWdlType).get //<-- here be `get` dragons
+        OutputDefinition(FullyQualifiedName(cop_id).id, womType, CommandOutputExpression(outputBinding, womType, inputNames))
+    }.toList
+
+    val inputs: List[_ <: Callable.InputDefinition] =
+      this.inputs.map { cip =>
+        val tpe = cip.`type`.flatMap(_.select[CwlType]).map(cwlTypeToWdlType).get
+
+        RequiredInputDefinition(FullyQualifiedName(cip.id).id, tpe)
+      }.toList
+
+    TaskDefinition(
+      id,
+      commandTemplate,
+      runtimeAttributes,
+      meta,
+      parameterMeta,
+      outputs,
+      inputs,
+      // TODO: This doesn't work in all cases and it feels clunky anyway - find a way to sort that out
+      prefixSeparator = "#",
+      commandPartSeparator = " "
+    )
+  }
+
+  def asCwl = Coproduct[Cwl](this)
+}
+
+object CommandLineTool {
+
+  def apply(inputs: Array[CommandInputParameter] = Array.empty,
+            outputs: Array[CommandOutputParameter] = Array.empty,
+            id: String,
+            requirements: Option[Array[Requirement]] = None,
+            hints: Option[Array[Map[String, String]]] = None,
+            label: Option[String] = None,
+            doc: Option[String] = None,
+            cwlVersion: Option[CwlVersion] = Option(CwlVersion.Version1),
+            baseCommand: Option[BaseCommand] = None,
+            arguments: Option[Array[CommandLineTool.Argument]] = None,
+            stdin: Option[StringOrExpression] = None,
+            stderr: Option[StringOrExpression] = None,
+            stdout: Option[StringOrExpression] = None,
+            successCodes: Option[Array[Int]] = None,
+            temporaryFailCodes: Option[Array[Int]] = None,
+            permanentFailCodes: Option[Array[Int]] = None): CommandLineTool  =
+              CommandLineTool(inputs, outputs, "CommandLineTool".narrow, id, requirements, hints, label, doc, cwlVersion, baseCommand, arguments, stdin, stderr, stdout, successCodes, temporaryFailCodes, permanentFailCodes)
+
+  type BaseCommand = String :+: Array[String] :+: CNil
+
+  type Argument = Expression :+: CommandLineBinding :+: String :+: CNil
+
+  case class CommandInputParameter(
+                                    id: String,
+                                    label: Option[String] = None,
+                                    secondaryFiles: Option[Array[Expression :+: String :+: CNil]] = None,
+                                    format: Option[Expression :+: Array[String] :+: String :+: CNil] = None, //only valid when type: File
+                                    streamable: Option[Boolean] = None, //only valid when type: File
+                                    doc: Option[String :+: Array[String] :+: CNil] = None,
+                                    inputBinding: Option[CommandLineBinding] = None,
+                                    default: Option[CwlAny] = None,
+                                    `type`: Option[MyriadInputType] = None)
+
+  case class CommandInputRecordSchema(
+                                       `type`: W.`"record"`.T,
+                                       fields: Option[Array[CommandInputRecordField]],
+                                       label: Option[String])
+
+  case class CommandInputRecordField(
+                                      name: String,
+                                      `type`: MyriadInputType,
+                                      doc: Option[String],
+                                      inputBinding: Option[CommandLineBinding],
+                                      label: Option[String])
+
+  case class CommandInputEnumSchema(
+                                     symbols: Array[String],
+                                     `type`: W.`"enum"`.T,
+                                     label: Option[String],
+                                     inputBinding: Option[CommandLineBinding])
+
+  case class CommandInputArraySchema(
+                                      items:
                                       CwlType :+:
-                                      CommandInputRecordSchema :+:
-                                      CommandInputEnumSchema :+:
-                                      CommandInputArraySchema :+:
-                                      String :+:
-                                      Array[
-                                        CwlType :+:
                                         CommandInputRecordSchema :+:
                                         CommandInputEnumSchema :+:
                                         CommandInputArraySchema :+:
                                         String :+:
-                                        CNil] :+:
-                                      CNil,
-                                    `type`: W.`"array"`.T,
-                                    label: Option[String],
-                                    inputBinding: Option[CommandLineBinding])
+                                        Array[
+                                          CwlType :+:
+                                            CommandInputRecordSchema :+:
+                                            CommandInputEnumSchema :+:
+                                            CommandInputArraySchema :+:
+                                            String :+:
+                                            CNil] :+:
+                                        CNil,
+                                      `type`: W.`"array"`.T,
+                                      label: Option[String],
+                                      inputBinding: Option[CommandLineBinding])
 
 
-case class CommandOutputParameter(
-                                   id: String,
-                                   label: Option[String] = None,
-                                   secondaryFiles: Option[Expression :+: String :+: Array[Expression :+: String :+: CNil] :+: CNil] = None,
-                                   format: Option[Expression :+: Array[String] :+: String :+: CNil] = None, //only valid when type: File
-                                   streamable: Option[Boolean] = None, //only valid when type: File
-                                   doc: Option[String :+: Array[String] :+: CNil] = None,
-                                   outputBinding: Option[CommandOutputBinding] = None,
-                                   `type`: Option[MyriadOutputType] = None)
+  case class CommandOutputParameter(
+                                     id: String,
+                                     label: Option[String] = None,
+                                     secondaryFiles: Option[Expression :+: String :+: Array[Expression :+: String :+: CNil] :+: CNil] = None,
+                                     format: Option[Expression :+: Array[String] :+: String :+: CNil] = None, //only valid when type: File
+                                     streamable: Option[Boolean] = None, //only valid when type: File
+                                     doc: Option[String :+: Array[String] :+: CNil] = None,
+                                     outputBinding: Option[CommandOutputBinding] = None,
+                                     `type`: Option[MyriadOutputType] = None)
 
 
+
+}
