--- cwl/src/main/scala/cwl/CwlFile.scala
@@ -1,157 +1,0 @@
-package cwl
-
-import common.Checked
-import common.validation.Checked._
-import shapeless.syntax.singleton._
-import shapeless.{:+:, CNil, Poly1, Witness, _}
-import CommandLineTool.BaseCommand
-import CwlType.CwlType
-import CwlVersion._
-import wom.callable.Callable.{OutputDefinition, RequiredInputDefinition}
-import wom.callable.{Callable, TaskDefinition}
-import wom.executable.Executable
-import wom.expression.WomExpression
-import wom.types.WomFileType
-import wom.{CommandPart, RuntimeAttributes}
-
-/**
-  *
-  * @param inputs
-  * @param outputs
-  * @param `class` This _should_ always be "CommandLineTool," however the spec does not -er- specify this.
-  * @param id
-  * @param requirements
-  * @param hints
-  * @param label
-  * @param doc
-  * @param cwlVersion
-  * @param baseCommand
-  * @param arguments
-  * @param stdin
-  * @param stderr
-  * @param stdout
-  * @param successCodes
-  * @param temporaryFailCodes
-  * @param permanentFailCodes
-  */
-case class CommandLineTool private(
-                                   inputs: Array[CommandInputParameter],
-                                   outputs: Array[CommandOutputParameter],
-                                   `class`: Witness.`"CommandLineTool"`.T,
-                                   id: String,
-                                   requirements: Option[Array[Requirement]],
-
-                                   //TODO: Fix this when CwlAny parses correctly
-                                   //hints: Option[Array[CwlAny]] = None,
-                                   hints: Option[Array[Map[String, String]]],
-
-                                   label: Option[String],
-                                   doc: Option[String],
-                                   cwlVersion: Option[CwlVersion],
-                                   baseCommand: Option[BaseCommand],
-                                   arguments: Option[Array[CommandLineTool.Argument]],
-                                   stdin: Option[StringOrExpression],
-                                   stderr: Option[StringOrExpression],
-                                   stdout: Option[StringOrExpression],
-                                   successCodes: Option[Array[Int]],
-                                   temporaryFailCodes: Option[Array[Int]],
-                                   permanentFailCodes: Option[Array[Int]]) {
-
-  def womExecutable(inputFile: Option[String] = None): Checked[Executable] =
-    CwlExecutableValidation.builWomExecutable(taskDefinition.validNelCheck, inputFile)
-
-  object ArgumentToId extends Poly1 {
-    implicit def ecmaScript:Case.Aux[Expression, String] = at[Expression] {
-      _ => ???
-    }
-
-    implicit def commandLineBinding = at[CommandLineBinding] { _ => "" }
-
-    implicit def string = at[String] {
-      identity
-    }
-  }
-
-  def taskDefinition: TaskDefinition = {
-
-    val id = this.id
-
-    val commandTemplate: Seq[CommandPart] = baseCommand.toSeq.flatMap(_.fold(BaseCommandToCommandParts)) ++
-      arguments.toSeq.flatMap(_.map(_.fold(ArgumentToCommandPart)))
-
-    val runtimeAttributes: RuntimeAttributes = RuntimeAttributes(Map.empty[String, WomExpression])
-
-    val meta: Map[String, String] = Map.empty
-    val parameterMeta: Map[String, String] = Map.empty
-
-    /*
-    quoted from: http://www.commonwl.org/v1.0/CommandLineTool.html#CommandOutputBinding :
-
-    For inputs and outputs, we only keep the variable name in the definition
-
-    The output parameter value is generated by applying these operations in the following order:
-
-      glob
-      loadContents
-      outputEval
-      secondaryFiles
-     */
-
-    val inputNames = this.inputs.map(_.id).toSet
-
-    val outputs: List[Callable.OutputDefinition] = this.outputs.map {
-      case CommandOutputParameter(cop_id, _, _, _, _, _, Some(outputBinding), Some(outputType)) if outputType.select[CwlType].contains(CwlType.File) =>
-        OutputDefinition(FullyQualifiedName(cop_id).id, WomFileType, CommandOutputExpression(outputBinding, WomFileType, inputNames))
-      case CommandOutputParameter(cop_id, _, _, _, _, _, Some(outputBinding), Some(tpe)) =>
-        val womType = tpe.select[CwlType].map(cwlTypeToWdlType).get //<-- here be `get` dragons
-        OutputDefinition(FullyQualifiedName(cop_id).id, womType, CommandOutputExpression(outputBinding, womType, inputNames))
-    }.toList
-
-    val inputs: List[_ <: Callable.InputDefinition] =
-      this.inputs.map { cip =>
-        val tpe = cip.`type`.flatMap(_.select[CwlType]).map(cwlTypeToWdlType).get
-
-        RequiredInputDefinition(FullyQualifiedName(cip.id).id, tpe)
-      }.toList
-
-    TaskDefinition(
-      id,
-      commandTemplate,
-      runtimeAttributes,
-      meta,
-      parameterMeta,
-      outputs,
-      inputs,
-      // TODO: This doesn't work in all cases and it feels clunky anyway - find a way to sort that out
-      prefixSeparator = "#",
-      commandPartSeparator = " "
-    )
-  }
-
-  def asCwl = Coproduct[Cwl](this)
-}
-
-object CommandLineTool {
-
-  def apply(inputs: Array[CommandInputParameter] = Array.empty,
-            outputs: Array[CommandOutputParameter] = Array.empty,
-            id: String,
-            requirements: Option[Array[Requirement]] = None,
-            hints: Option[Array[Map[String, String]]] = None,
-            label: Option[String] = None,
-            doc: Option[String] = None,
-            cwlVersion: Option[CwlVersion] = Option(CwlVersion.Version1),
-            baseCommand: Option[BaseCommand] = None,
-            arguments: Option[Array[CommandLineTool.Argument]] = None,
-            stdin: Option[StringOrExpression] = None,
-            stderr: Option[StringOrExpression] = None,
-            stdout: Option[StringOrExpression] = None,
-            successCodes: Option[Array[Int]] = None,
-            temporaryFailCodes: Option[Array[Int]] = None,
-            permanentFailCodes: Option[Array[Int]] = None): CommandLineTool  =
-              CommandLineTool(inputs, outputs, "CommandLineTool".narrow, id, requirements, hints, label, doc, cwlVersion, baseCommand, arguments, stdin, stderr, stdout, successCodes, temporaryFailCodes, permanentFailCodes)
-
-  type BaseCommand = String :+: Array[String] :+: CNil
-
-  type Argument = Expression :+: CommandLineBinding :+: String :+: CNil
-}
