+++ cwl/src/main/scala/cwl/Workflow.scala
@@ -27,7 +27,7 @@ case class Workflow private(
                      outputs: Array[WorkflowOutputParameter],
                      steps: Array[WorkflowStep]) {
   def womExecutable(inputFile: Option[String] = None): Checked[Executable] = {
-    CwlExecutableValidation.builWomExecutable(womDefinition, inputFile)
+    CwlExecutableValidation.buildWomExecutable(womDefinition, inputFile)
   }
 
   val fileNames: List[String] = steps.toList.flatMap(_.run.select[String].toList)
